requirements.txt
fastapi
uvicorn
SQLAlchemy
sqlalchemy_utils
pydantic_settings
requests
psycopg2-binary
PyJWT
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
cohere
pdfplumber
python-multipart
weaviate-client
langchain
streamlit
request
requests_toolbelt
pypdf 